split config.updateFromArgv() into getArgv() and config.updateFromArgv(argv)
also add constants.js
